move asp writer files to common directory for anyone to use
