+++ src/AutoRest.CSharp/LowLevel/AutoRest/DpgOutputLibrary.cs
@@ -5,10 +5,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using AutoRest.CSharp.Common.Input;
+using AutoRest.CSharp.Common.Output.Models.Types;
 using AutoRest.CSharp.Generation.Types;
 using AutoRest.CSharp.Input;
 using AutoRest.CSharp.Input.Source;
-using AutoRest.CSharp.LowLevel.Output;
 
 namespace AutoRest.CSharp.Output.Models.Types
 {
